pyproject.toml
[project]
description = ""
authors = [
    # Указывать почту не обязательно
    {name = "Nikita", email = ""},
    {name = "Alexey", email = ""},
    {name = "Kostia", email = ""},
    {name = "Aleksandr", email = ""},
]
dependencies = [
    "python-dotenv>=1.0.1",
    "aiogram==2.21",
    "emoji==2.10.1",
    "bleach>=6.1.0",
    "pillow>=10.2.0",
    "apscheduler>=3.10.4",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

#region    -- Dev Tools
[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.1.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "flake8>=7.0.0",
    "pylint>=3.0.3",
]

[tool.pylint.MASTER]
ignore-paths = [ ".venv/*" ]

[tool.pylint."MESSAGES CONTROL"]
disable = '''
missing-module-docstring,
missing-class-docstring,
missing-function-docstring
'''


[tool.black]
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.mypy]
disallow_untyped_defs = true
disallow_incomplete_defs = true
python_version = "3.11"
strict = true
show_error_codes = true
warn_unused_ignores = true

[tool.pdm.scripts]
start = "pdm run python -m src"
# Currently formaters make code worse, TODO: configure them properly
#format-black = "black src/ tests/"
#format-isort = "isort src tests"
#format = {composite = ["format-black", "format-isort"]}
lint-black = "black --check --diff src/ tests/"
lint-flake8 = "flake8 src/ tests/"
lint-isort = "isort --check --diff src tests"
lint-mypy = "mypy src/ tests"
lint = {composite = ["lint-black", "lint-flake8", "lint-isort", "lint-mypy", "pylint --recursive=y ."]}
#endregion -- Dev Tools
